+++ src/main/java/com/itep/restaurant_service/services/impl/MenuServiceImpl.java
@@ -10,9 +10,12 @@ import com.itep.restaurant_service.repositories.entities.RestaurantEntity;
 import com.itep.restaurant_service.services.MenuService;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -40,24 +43,34 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     public MenuResource createMenu(Long rest_id, Long cat_id, MenuEntity body) throws Exception {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        // Get the username of the current user
+        String username = authentication.getName();
         Optional<CategoryEntity> yourEntityOptional = categoryRepository.findById(cat_id);
         if (yourEntityOptional .isPresent()) {
             CategoryEntity yourEntity = yourEntityOptional.get();
             if(yourEntity.getRestaurant().getId() == rest_id){
-                try{
-                    body.setCategory(yourEntity);
-                    return menuRepository.save(body).toMenuResource();
-                }
-                catch (Exception e){
-                    if(e.getMessage().contains("duplicate key value violates unique constraint")){
-                        throw new Exception("menu with this name already exists in category");
+                if(Objects.equals(yourEntity.getRestaurant().getOwner().getUsername(), username)){
+                    try{
+                        body.setCategory(yourEntity);
+                        return menuRepository.save(body).toMenuResource();
                     }
-                    else if (e.getMessage().contains("not-null property references a null")) {
-                        throw new Exception("You must provide all the category fields");
+                    catch (Exception e){
+                        if(e.getMessage().contains("duplicate key value violates unique constraint")){
+                            throw new Exception("menu with this name already exists in category");
+                        }
+                        else if (e.getMessage().contains("not-null property references a null")) {
+                            throw new Exception("You must provide all the category fields");
+                        }
+                        System.out.println(e.getMessage());
+                        throw new Exception("unknown error");
                     }
-                    System.out.println(e.getMessage());
-                    throw new Exception("unknown error");
                 }
+                else {
+                    throw new EntityNotFoundException("you are not the owner of Restaurant");
+                }
+
             }
             else{
                 throw new EntityNotFoundException("Category with id " + cat_id + " not belong to Restaurant");
@@ -72,6 +85,10 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     public MenuResource updateMenu(Long rest_id, Long cat_id,Long id, MenuEntity body) throws Exception {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        // Get the username of the current user
+        String username = authentication.getName();
         Optional<CategoryEntity> yourEntityOptional = categoryRepository.findById(cat_id);
         if (yourEntityOptional .isPresent()) {
             CategoryEntity yourEntity = yourEntityOptional.get();
@@ -79,13 +96,18 @@ public class MenuServiceImpl implements MenuService {
                 MenuEntity existMenu = menuRepository.findById(id)
                         .orElseThrow(() -> new Exception("Menu not found"));
                 if(existMenu.getCategory().getId() == cat_id){
-
-                    try {
-                        existMenu.setName(body.getName());
-                        return menuRepository.save(existMenu).toMenuResource();
-                    } catch (Exception e) {
-                        throw new Exception("Failed to update menu");
+                    if(Objects.equals(yourEntity.getRestaurant().getOwner().getUsername(), username)){
+                        try {
+                            existMenu.setName(body.getName());
+                            return menuRepository.save(existMenu).toMenuResource();
+                        } catch (Exception e) {
+                            throw new Exception("Failed to update menu");
+                        }
                     }
+                    else {
+                        throw new EntityNotFoundException("you are not the owner of Restaurant");
+                    }
+
                 }
                 else{
                     throw new EntityNotFoundException("Menu with id " + id + " not belong to Category");
@@ -105,6 +127,10 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     public void deleteMenu(Long rest_id, Long cat_id,Long id) throws Exception {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        // Get the username of the current user
+        String username = authentication.getName();
         Optional<CategoryEntity> yourEntityOptional = categoryRepository.findById(cat_id);
         if (yourEntityOptional .isPresent()) {
             CategoryEntity yourEntity = yourEntityOptional.get();
@@ -112,11 +138,17 @@ public class MenuServiceImpl implements MenuService {
                 MenuEntity existMenu = menuRepository.findById(id)
                         .orElseThrow(() -> new Exception("Menu not found"));
                 if(existMenu.getCategory().getId() == cat_id){
-                    try {
-                        menuRepository.deleteById(id);
-                    } catch (Exception e) {
-                        throw new Exception("Failed to update menu");
+                    if(Objects.equals(yourEntity.getRestaurant().getOwner().getUsername(), username)){
+                        try {
+                            menuRepository.deleteById(id);
+                        } catch (Exception e) {
+                            throw new Exception("Failed to update menu");
+                        }
                     }
+                    else {
+                        throw new EntityNotFoundException("you are not the owner of Restaurant");
+                    }
+
                 }
                 else{
                     throw new EntityNotFoundException("Menu with id " + id + " not belong to Category");
